fix config loading
