--- src/platform/linux/config.h
@@ -1,7 +1,0 @@
-
-#ifndef _GNEISS_CONFIG_H_
-#define _GNEISS_CONFIG_H_
-
-void gneiss_load_config(const char *file, char **content);
-
-#endif /* ifndef _GNEISS_CONFIG_H_ */
